review: break import cycle + drop redundant pre-delete read
Addresses two cleanup items on PR #2263.

1. Circular import between resourceReview.service.ts and user.service.ts
   was a webpack/Next.js footgun even though it didn't fail at runtime
   (the cycle only fires through the cross-service-imported symbol
   inside a runtime function body). Extracted bustRatingTotalsCache and
   bustRatingTotalsForRows into a new standalone module
   src/server/services/resourceReview.cache.ts. Both services now import
   from there — no cycle.

2. deleteResourceReview did an extra dbRead.findUnique to grab modelId
   and modelVersionId before the delete. Prisma's .delete() already
   returns the full row by default (the controller in
   resourceReview.controller.ts:120-128 already relies on this), so the
   pre-fetch was a wasted round-trip. Use ret.modelId /
   ret.modelVersionId from the delete result.
